API

Discorder.AbstractEventPublisher

Discorder.AbstractTrigger

Discorder.CommandTrigger

Discorder.Event

Discorder.GatewayTracker

Discorder.ReactionAddTrigger

Discorder.add_group_dm_recipient

Discorder.add_guild_member

Discorder.add_guild_member_role

Discorder.add_thread_member

Discorder.begin_guild_prune

Discorder.bulk_delete_messages

Discorder.create_channel_invite

Discorder.create_dm

Discorder.create_event

Discorder.create_group_dm

Discorder.create_guild

Discorder.create_guild_ban

Discorder.create_guild_channel

Discorder.create_guild_emoji

Discorder.create_guild_from_guild_template

Discorder.create_guild_role

Discorder.create_guild_scheduled_event

Discorder.create_guild_template

Discorder.create_guild_ticker

Discorder.create_message

Discorder.create_reaction

Discorder.create_stage_instance

Discorder.create_webhook

Discorder.crosspost_message

Discorder.delete_all_reactions

Discorder.delete_all_reactions_for_emoji

Discorder.delete_channel

Discorder.delete_channel_permission

Discorder.delete_guild

Discorder.delete_guild_emoji

Discorder.delete_guild_integration

Discorder.delete_guild_role

Discorder.delete_guild_scheduled_event

Discorder.delete_guild_sticker

Discorder.delete_guild_template

Discorder.delete_invite

Discorder.delete_message

Discorder.delete_own_reaction

Discorder.delete_stage_instance

Discorder.delete_user_reaction

Discorder.delete_webhook

Discorder.delete_webhook_message

Discorder.delete_webhook_with_token

Discorder.edit_channel_permissions

Discorder.edit_message

Discorder.edit_webhook_message

Discorder.execute_github_compatible_webhook

Discorder.execute_slack_compatible_webhook

Discorder.execute_webhook

Discorder.follow_news_channel

Discorder.get_bot_token

Discorder.get_channel

Discorder.get_channel_invites

Discorder.get_channel_message

Discorder.get_channel_messages

Discorder.get_channel_webhooks

Discorder.get_current_user

Discorder.get_current_user_guild_member

Discorder.get_current_user_guilds

Discorder.get_gateway

Discorder.get_guild

Discorder.get_guild_audit_log

Discorder.get_guild_ban

Discorder.get_guild_bans

Discorder.get_guild_channels

Discorder.get_guild_emoji

Discorder.get_guild_integrations

Discorder.get_guild_invites

Discorder.get_guild_member

Discorder.get_guild_preview

Discorder.get_guild_prune_count

Discorder.get_guild_roles

Discorder.get_guild_scheduled_event

Discorder.get_guild_scheduled_event_users

Discorder.get_guild_sticker

Discorder.get_guild_template

Discorder.get_guild_templates

Discorder.get_guild_vanity_url

Discorder.get_guild_voice_regions

Discorder.get_guild_webhooks

Discorder.get_guild_welcome_screen

Discorder.get_guild_widget

Discorder.get_guild_widget_image

Discorder.get_guild_widget_settings

Discorder.get_invite

Discorder.get_logger

Discorder.get_pinned_messages

Discorder.get_reactions

Discorder.get_stage_instance

Discorder.get_sticker

Discorder.get_thread_member

Discorder.get_user

Discorder.get_user_connections

Discorder.get_webhook

Discorder.get_webhook_message

Discorder.get_webhook_with_token

Discorder.join_thread

Discorder.json

Discorder.leave_guild

Discorder.leave_thread

Discorder.list_active_guild_threads

Discorder.list_guild_emojis

Discorder.list_guild_members

Discorder.list_guild_scheduled_events

Discorder.list_guild_stickers

Discorder.list_joined_private_archived_threads

Discorder.list_nitro_sticker_packs

Discorder.list_private_archived_threads

Discorder.list_public_archived_threads

Discorder.list_thread_members

Discorder.list_voice_regions

Discorder.modify_channel

Discorder.modify_current_member

Discorder.modify_current_user

Discorder.modify_current_user_nick

Discorder.modify_current_user_voice_state

Discorder.modify_guild

Discorder.modify_guild_channel_positions

Discorder.modify_guild_emoji

Discorder.modify_guild_member

Discorder.modify_guild_role

Discorder.modify_guild_role_positions

Discorder.modify_guild_scheduled_event

Discorder.modify_guild_sticker

Discorder.modify_guild_template

Discorder.modify_guild_welcome_screen

Discorder.modify_guild_widget

Discorder.modify_stage_instance

Discorder.modify_user_voice_state

Discorder.modify_webhook

Discorder.modify_webhook_with_token

Discorder.pin_message

Discorder.publish

Discorder.remove_group_dm_recipient

Discorder.remove_guild_ban

Discorder.remove_guild_member

Discorder.remove_guild_member_role

Discorder.remove_thread_member

Discorder.safe_parse_json

Discorder.safe_wait

Discorder.sanitize

Discorder.search_guild_members

Discorder.serve

Discorder.should_trigger

Discorder.show_error

Discorder.start_control_plane

Discorder.start_thread_from_message

Discorder.start_thread_without_message

Discorder.sync_guild_template

Discorder.trigger_typing_indicator

Discorder.unpin_message

Discorder.waterfall

Discorder.@discord_object

Discorder.AbstractEventPublisher — Type

AbstractEventPublisher

An interface for publishing gateway events.

Discorder.AbstractTrigger — Type

AbstractTrigger

A trigger is used to represent data that can be used to determine whether an event handler should be fired. For example, a regex-based command trigger may store a Regex object that can be used to match messages.

Discorder.CommandTrigger — Type

CommandTrigger

A regex-based trigger that matches a message pattern.

Discorder.Event — Type

Event

A gateway event.

Discorder.GatewayTracker — Type

GatewayTracker

GatewayTracker is a stateful object used by the Control Pane.

Discorder.ReactionAddTrigger — Type

ReactionAddTrigger

A trigger that is fired when a reaction is added to a message.

Discorder.add_group_dm_recipient — Function

add_group_dm_recipient(client, channel_id, user_id; kwargs...) -> Nothing

Make a PUT request to /channels/<channel_id>/recipients/<user_id>. See the Discord API documentation for more information.

Discorder.add_guild_member — Function

add_guild_member(client, guild_id, user_id; kwargs...) -> GuildMember

Make a PUT request to /guilds/<guild_id>/members/<user_id>. See the Discord API documentation for more information.

Discorder.add_guild_member_role — Function

add_guild_member_role(client, guild_id, user_id, role_id) -> Nothing

Make a PUT request to /guilds/<guild_id>/members/<user_id>/roles/<role_id>. See the Discord API documentation for more information.

Discorder.add_thread_member — Function

add_thread_member(client, channel_id, user_id) -> Nothing

Make a PUT request to /channels/<channel_id>/thread-members/<user_id>. See the Discord API documentation for more information.

Discorder.begin_guild_prune — Function

begin_guild_prune(client, guild_id; kwargs...) -> PruneCount

Make a POST request to /guilds/<guild_id>/prune. See the Discord API documentation for more information.

Discorder.bulk_delete_messages — Function

bulk_delete_messages(client, channel_id; kwargs...) -> Nothing

Make a POST request to /channels/<channel_id>/messages/bulk-delete. See the Discord API documentation for more information.

Discorder.create_channel_invite — Function

create_channel_invite(client, channel_id; kwargs...) -> Invite

Make a POST request to /channels/<channel_id>/invites. See the Discord API documentation for more information.

Discorder.create_dm — Function

create_dm(client; kwargs...) -> DiscordChannel

Make a POST request to /users/@me/channels. See the Discord API documentation for more information.

Discorder.create_event — Method

create_event_object(tracker::GatewayTracker, event_type::AbstractString, json::Optional{Abstract)

Create an event object by parsing the gateway event as JSON string.

Discorder.create_group_dm — Function

create_group_dm(client; kwargs...) -> DiscordChannel

Make a POST request to /users/@me/channels. See the Discord API documentation for more information.

Discorder.create_guild — Function

create_guild(client; kwargs...) -> Guild

Make a POST request to /guilds. See the Discord API documentation for more information.

Discorder.create_guild_ban — Function

create_guild_ban(client, guild_id, user_id; kwargs...) -> Nothing

Make a PUT request to /guilds/<guild_id>/bans/<user_id>. See the Discord API documentation for more information.

Discorder.create_guild_channel — Function

create_guild_channel(client, guild_id; kwargs...) -> DiscordChannel

Make a POST request to /guilds/<guild_id>/channels. See the Discord API documentation for more information.

Discorder.create_guild_emoji — Function

create_guild_emoji(client, guild_id; kwargs...) -> Emoji

Make a POST request to /guilds/<guild_id>/emojis. See the Discord API documentation for more information.

Discorder.create_guild_from_guild_template — Function

create_guild_from_guild_template(client, template_code; kwargs...) -> Guild

Make a POST request to /guilds/templates/<template_code>. See the Discord API documentation for more information.

Discorder.create_guild_role — Function

create_guild_role(client, guild_id; kwargs...) -> Role

Make a POST request to /guilds/<guild_id>/roles. See the Discord API documentation for more information.

Discorder.create_guild_scheduled_event — Function

create_guild_scheduled_event(client, guild_id; kwargs...) -> GuildScheduledEvent

Make a POST request to /guilds/<guild_id>/scheduled-events. See the Discord API documentation for more information.

Discorder.create_guild_template — Function

create_guild_template(client, guild_id; kwargs...) -> GuildTemplate

Make a POST request to /guilds/<guild_id>/templates. See the Discord API documentation for more information.

Discorder.create_guild_ticker — Function

create_guild_ticker(client, guild_id; kwargs...) -> Sticker

Make a POST request to /guilds/<guild_id>/stickers. See the Discord API documentation for more information.

Discorder.create_message — Function

create_message(client, channel_id; kwargs...) -> Message

Make a POST request to /channels/<channel_id>/messages. See the Discord API documentation for more information.

Discorder.create_reaction — Function

create_reaction(client, channel_id, message_id, emoji) -> Nothing

Make a PUT request to /channels/<channel_id>/messages/<message_id>/reactions/<emoji>/@me. See the Discord API documentation for more information.

Discorder.create_stage_instance — Function

create_stage_instance(client; kwargs...) -> StageInstance

Make a POST request to /stage-instances. See the Discord API documentation for more information.

Discorder.create_webhook — Function

create_webhook(client, channel_id; kwargs...) -> Webhook

Make a POST request to /channels/<channel_id>/webhooks. See the Discord API documentation for more information.

Discorder.crosspost_message — Function

crosspost_message(client, channel_id, message_id) -> Message

Make a POST request to /channels/<channel_id>/messages/<message_id>/crosspost. See the Discord API documentation for more information.

Discorder.delete_all_reactions — Function

delete_all_reactions(client, channel_id, message_id) -> Nothing

Make a DELETE request to /channels/<channel_id>/messages/<message_id>/reactions. See the Discord API documentation for more information.

Discorder.delete_all_reactions_for_emoji — Function

delete_all_reactions_for_emoji(client, channel_id, message_id, emoji) -> Nothing

Make a DELETE request to /channels/<channel_id>/messages/<message_id>/reactions/<emoji>. See the Discord API documentation for more information.

Discorder.delete_channel — Function

delete_channel(client, channel_id) -> DiscordChannel

Make a DELETE request to /channels/<channel_id>. See the Discord API documentation for more information.

Discorder.delete_channel_permission — Function

delete_channel_permission(client, channel_id, overwrite) -> Nothing

Make a DELETE request to /channels/<channel_id>/permissions/<overwrite>. See the Discord API documentation for more information.

Discorder.delete_guild — Function

delete_guild(client, guild_id) -> Nothing

Make a DELETE request to /guilds/<guild_id>. See the Discord API documentation for more information.

Discorder.delete_guild_emoji — Function

delete_guild_emoji(client, guild_id, emoji) -> Nothing

Make a DELETE request to /guilds/<guild_id>/emojis/<emoji>. See the Discord API documentation for more information.

Discorder.delete_guild_integration — Function

delete_guild_integration(client, guild_id, integration_id) -> Nothing

Make a DELETE request to /guilds/<guild_id>/integrations/<integration_id>. See the Discord API documentation for more information.

Discorder.delete_guild_role — Function

delete_guild_role(client, guild_id, role_id) -> Nothing

Make a DELETE request to /guilds/<guild_id>/roles/<role_id>. See the Discord API documentation for more information.

Discorder.delete_guild_scheduled_event — Function

delete_guild_scheduled_event(client, guild_id, guild_scheduled_event_id) -> Nothing

Make a DELETE request to /guilds/<guild_id>/scheduled-events/<guild_scheduled_event_id>. See the Discord API documentation for more information.

Discorder.delete_guild_sticker — Function

delete_guild_sticker(client, guild_id, sticker_id) -> Nothing

Make a DELETE request to /guilds/<guild_id>/stickers/<sticker_id>. See the Discord API documentation for more information.

Discorder.delete_guild_template — Function

delete_guild_template(client, guild_id, template_code) -> GuildTemplate

Make a DELETE request to /guilds/<guild_id>/templates/<template_code>. See the Discord API documentation for more information.

Discorder.delete_invite — Function

delete_invite(client, invite_code) -> Invite

Make a DELETE request to /invites/<invite_code>. See the Discord API documentation for more information.

Discorder.delete_message — Function

delete_message(client, channel_id, message_id) -> Nothing

Make a DELETE request to /channels/<channel_id>/messages/<message_id>. See the Discord API documentation for more information.

Discorder.delete_own_reaction — Function

delete_own_reaction(client, channel_id, message_id, emoji) -> Nothing

Make a DELETE request to /channels/<channel_id>/messages/<message_id>/reactions/<emoji>/@me. See the Discord API documentation for more information.

Discorder.delete_stage_instance — Function

delete_stage_instance(client) -> Nothing

Make a DELETE request to /stage-instances/{channel.id}. See the Discord API documentation for more information.

Discorder.delete_user_reaction — Function

delete_user_reaction(client, channel_id, message_id, emoji, user_id) -> Nothing

Make a DELETE request to /channels/<channel_id>/messages/<message_id>/reactions/<emoji>/<user_id>. See the Discord API documentation for more information.

Discorder.delete_webhook — Function

delete_webhook(client, webhook_id; kwargs...) -> Webhook

Make a DELETE request to /webhooks/<webhook_id>. See the Discord API documentation for more information.

Discorder.delete_webhook_message — Function

delete_webhook_message(client, webhook_id, webhook_tokenb, message_id; kwargs...) -> Nothing

Make a DELETE request to /webhooks/<webhook_id>/<webhook_tokenb>/messages/<message_id>. See the Discord API documentation for more information.

Discorder.delete_webhook_with_token — Function

delete_webhook_with_token(client, webhook_id, webhook_token; kwargs...) -> Webhook

Make a DELETE request to /webhooks/<webhook_id>/<webhook_token>. See the Discord API documentation for more information.

Discorder.edit_channel_permissions — Function

edit_channel_permissions(client, channel_id, overwrite; kwargs...) -> Nothing

Make a PUT request to /channels/<channel_id>/permissions/<overwrite>. See the Discord API documentation for more information.

Discorder.edit_message — Function

edit_message(client, channel_id, message_id; kwargs...) -> Message

Make a PATCH request to /channels/<channel_id>/messages/<message_id>. See the Discord API documentation for more information.

Discorder.edit_webhook_message — Function

edit_webhook_message(client, webhook_id, webhook_tokenb, message_id; kwargs...) -> Message

Make a PATCH request to /webhooks/<webhook_id>/<webhook_tokenb>/messages/<message_id>. See the Discord API documentation for more information.

Discorder.execute_github_compatible_webhook — Function

execute_github_compatible_webhook(client, webhook_id, webhook_token; wait=true, kwargs...) -> Message

Make a POST request to /webhooks/<webhook_id>/<webhook_token>/github. See the Discord API documentation for more information.

Discorder.execute_slack_compatible_webhook — Function

execute_slack_compatible_webhook(client, webhook_id, webhook_token; wait=true, kwargs...) -> Message

Make a POST request to /webhooks/<webhook_id>/<webhook_token>/slack. See the Discord API documentation for more information.

Discorder.execute_webhook — Function

execute_webhook(client, webhook_id, webhook_token; wait=true, kwargs...) -> Message

Make a POST request to /webhooks/<webhook_id>/<webhook_token>. See the Discord API documentation for more information.

Discorder.follow_news_channel — Function

follow_news_channel(client, channel_id; kwargs...) -> FollowedChannel

Make a POST request to /channels/<channel_id>/followers. See the Discord API documentation for more information.

Discorder.get_bot_token — Method

get_bot_token()

Get the bot token from enviornment variable DISCORD_BOT_TOKEN.

Discorder.get_channel — Function

get_channel(client, channel_id; kwargs...) -> DiscordChannel

Make a GET request to /channels/<channel_id>. See the Discord API documentation for more information.

Discorder.get_channel_invites — Function

get_channel_invites(client, channel_id; kwargs...) -> Vector{Invite}

Make a GET request to /channels/<channel_id>/invites. See the Discord API documentation for more information.

Discorder.get_channel_message — Function

get_channel_message(client, channel_id, message_id) -> Message

Make a GET request to /channels/<channel_id>/messages/<message_id>. See the Discord API documentation for more information.

Discorder.get_channel_messages — Function

get_channel_messages(client, channel_id; kwargs...) -> Vector{Message}

Make a GET request to /channels/<channel_id>/messages. See the Discord API documentation for more information.

Discorder.get_channel_webhooks — Function

get_channel_webhooks(client, channel_id) -> Vector{Webhook}

Make a GET request to /channels/<channel_id>/webhooks. See the Discord API documentation for more information.

Discorder.get_current_user — Function

get_current_user(client) -> User

Make a GET request to /users/@me. See the Discord API documentation for more information.

Discorder.get_current_user_guild_member — Function

get_current_user_guild_member(client) -> GuildMember

Make a GET request to /users/@me/guilds/{guild.id}/member. See the Discord API documentation for more information.

Discorder.get_current_user_guilds — Function

get_current_user_guilds(client; kwargs...) -> Vector{Guild}

Make a GET request to /users/@me/guilds. See the Discord API documentation for more information.

Discorder.get_gateway — Function

get_gateway(client) -> Gateway

Make a GET request to /gateway. See the Discord API documentation for more information.

Discorder.get_guild — Function

get_guild(client, guild_id; kwargs...) -> Guild

Make a GET request to /guilds/<guild_id>. See the Discord API documentation for more information.

Discorder.get_guild_audit_log — Function

get_guild_audit_log(client, guild_id; kwargs...) -> AuditLog

Make a GET request to /guilds/<guild_id>/audit-logs. See the Discord API documentation for more information.

Discorder.get_guild_ban — Function

get_guild_ban(client, guild_id, user_id) -> Ban

Make a GET request to /guilds/<guild_id>/bans/<user_id>. See the Discord API documentation for more information.

Discorder.get_guild_bans — Function

get_guild_bans(client, guild_id) -> Vector{Guild}

Make a GET request to /guilds/<guild_id>/bans. See the Discord API documentation for more information.

Discorder.get_guild_channels — Function

get_guild_channels(client, guild_id) -> Vector{DiscordChannel}

Make a GET request to /guilds/<guild_id>/channels. See the Discord API documentation for more information.

Discorder.get_guild_emoji — Function

get_guild_emoji(client, guild_id, emoji) -> Emoji

Make a GET request to /guilds/<guild_id>/emojis/<emoji>. See the Discord API documentation for more information.

Discorder.get_guild_integrations — Function

get_guild_integrations(client, guild_id) -> Vector{Integration}

Make a GET request to /guilds/<guild_id>/integrations. See the Discord API documentation for more information.

Discorder.get_guild_invites — Function

get_guild_invites(client, guild_id) -> Vector{Invite}

Make a GET request to /guilds/<guild_id>/invites. See the Discord API documentation for more information.

Discorder.get_guild_member — Function

get_guild_member(client, guild_id, user_id) -> GuildMember

Make a GET request to /guilds/<guild_id>/members/<user_id>. See the Discord API documentation for more information.

Discorder.get_guild_preview — Function

get_guild_preview(client, guild_id) -> Guild

Make a GET request to /guilds/<guild_id>/preview. See the Discord API documentation for more information.

Discorder.get_guild_prune_count — Function

get_guild_prune_count(client, guild_id; kwargs...) -> PruneCount

Make a GET request to /guilds/<guild_id>/prune. See the Discord API documentation for more information.

Discorder.get_guild_roles — Function

get_guild_roles(client, guild_id) -> Vector{Role}

Make a GET request to /guilds/<guild_id>/roles. See the Discord API documentation for more information.

Discorder.get_guild_scheduled_event — Function

get_guild_scheduled_event(client, guild_id, guild_scheduled_event_id; kwargs...) -> GuildScheduledEvent

Make a GET request to /guilds/<guild_id>/scheduled-events/<guild_scheduled_event_id>. See the Discord API documentation for more information.

Discorder.get_guild_scheduled_event_users — Function

get_guild_scheduled_event_users(client, guild_id, guild_scheduled_event_id; kwargs...) -> Vector{GuildScheduledEventUser}

Make a GET request to /guilds/<guild_id>/scheduled-events/<guild_scheduled_event_id>/users. See the Discord API documentation for more information.

Discorder.get_guild_sticker — Function

get_guild_sticker(client, guild_id, sticker_id) -> Sticker

Make a GET request to /guilds/<guild_id>/stickers/<sticker_id>. See the Discord API documentation for more information.

Discorder.get_guild_template — Function

get_guild_template(client, template_code) -> GuildTemplate

Make a GET request to /guilds/templates/<template_code>. See the Discord API documentation for more information.

Discorder.get_guild_templates — Function

get_guild_templates(client, guild_id) -> Vector{GuildTemplate}

Make a GET request to /guilds/<guild_id>/templates. See the Discord API documentation for more information.

Discorder.get_guild_vanity_url — Function

get_guild_vanity_url(client, guild_id) -> Invite

Make a GET request to /guilds/<guild_id>/vanity-url. See the Discord API documentation for more information.

Discorder.get_guild_voice_regions — Function

get_guild_voice_regions(client, guild_id) -> Vector{VoiceRegion}

Make a GET request to /guilds/<guild_id>/regions. See the Discord API documentation for more information.

Discorder.get_guild_webhooks — Function

get_guild_webhooks(client, guild_id) -> Vector{Webhook}

Make a GET request to /guilds/<guild_id>/webhooks. See the Discord API documentation for more information.

Discorder.get_guild_welcome_screen — Function

get_guild_welcome_screen(client, guild_id; kwargs...) -> WelcomeScreen

Make a GET request to /guilds/<guild_id>/welcome-screen. See the Discord API documentation for more information.

Discorder.get_guild_widget — Function

get_guild_widget(client, guild_id) -> GuildWidget

Make a GET request to /guilds/<guild_id>/widget.json. See the Discord API documentation for more information.

Discorder.get_guild_widget_image — Function

get_guild_widget_image(client, guild_id) -> String

Make a GET request to /guilds/<guild_id>/widget.png. See the Discord API documentation for more information.

Discorder.get_guild_widget_settings — Function

get_guild_widget_settings(client, guild_id) -> GuildWidgetSettings

Make a GET request to /guilds/<guild_id>/widget. See the Discord API documentation for more information.

Discorder.get_invite — Function

get_invite(client, invite_code; kwargs...) -> Invite

Make a GET request to /invites/<invite_code>. See the Discord API documentation for more information.

Discorder.get_logger — Method

get_logger(filename; debug::Bool)

Return a custom logger that write to the specified log file. A timestamp is automatically injected.

Arguments

file_path: path of log file

Keyword arguments

debug: turn on debug logging (default = false)

Discorder.get_pinned_messages — Function

get_pinned_messages(client, channel_id) -> Vector{Message}

Make a GET request to /channels/<channel_id>/pins. See the Discord API documentation for more information.

Discorder.get_reactions — Function

get_reactions(client, channel_id, message_id, emoji; kwargs...) -> Vector{User}

Make a GET request to /channels/<channel_id>/messages/<message_id>/reactions/<emoji>. See the Discord API documentation for more information.

Discorder.get_stage_instance — Function

get_stage_instance(client) -> StageInstance

Make a GET request to /stage-instances/{channel.id}. See the Discord API documentation for more information.

Discorder.get_sticker — Function

get_sticker(client, sticker_id) -> Sticker

Make a GET request to /stickers/<sticker_id>. See the Discord API documentation for more information.

Discorder.get_thread_member — Function

get_thread_member(client, channel_id, user_id) -> ThreadMember

Make a GET request to /channels/<channel_id>/thread-members/<user_id>. See the Discord API documentation for more information.

Discorder.get_user — Function

get_user(client, user_id) -> User

Make a GET request to /users/<user_id>. See the Discord API documentation for more information.

Discorder.get_user_connections — Function

get_user_connections(client) -> Vector{Connection}

Make a GET request to /users/@me/connections. See the Discord API documentation for more information.

Discorder.get_webhook — Function

get_webhook(client, webhook_id) -> Webhook

Make a GET request to /webhooks/<webhook_id>. See the Discord API documentation for more information.

Discorder.get_webhook_message — Function

get_webhook_message(client, webhook_id, webhook_tokenb, message_id; kwargs...) -> Message

Make a GET request to /webhooks/<webhook_id>/<webhook_tokenb>/messages/<message_id>. See the Discord API documentation for more information.

Discorder.get_webhook_with_token — Function

get_webhook_with_token(client, webhook_id, webhook_token) -> Webhook

Make a GET request to /webhooks/<webhook_id>/<webhook_token>. See the Discord API documentation for more information.

Discorder.join_thread — Function

join_thread(client, channel_id) -> Nothing

Make a PUT request to /channels/<channel_id>/thread-members/@me. See the Discord API documentation for more information.

Discorder.json — Method

Convert object to JSON formatted string.

Fields with missing values are excluded.

Fields with nothing values are included as null.

Discorder.leave_guild — Function

leave_guild(client, guild_id) -> Nothing

Make a DELETE request to /users/@me/guilds/<guild_id>. See the Discord API documentation for more information.

Discorder.leave_thread — Function

leave_thread(client, channel_id) -> Nothing

Make a DELETE request to /channels/<channel_id>/thread-members/@me. See the Discord API documentation for more information.

Discorder.list_active_guild_threads — Function

list_active_guild_threads(client, guild_id) -> Nothing

Make a GET request to /guilds/<guild_id>/threads/active. See the Discord API documentation for more information.

Discorder.list_guild_emojis — Function

list_guild_emojis(client, guild_id) -> Vector{Emoji}

Make a GET request to /guilds/<guild_id>/emojis. See the Discord API documentation for more information.

Discorder.list_guild_members — Function

list_guild_members(client, guild_id; kwargs...) -> Vector{GuildMember}

Make a GET request to /guilds/<guild_id>/members. See the Discord API documentation for more information.

Discorder.list_guild_scheduled_events — Function

list_guild_scheduled_events(client, guild_id; kwargs...) -> Vector{GuildScheduledEvent}

Make a GET request to /guilds/<guild_id>/scheduled-events. See the Discord API documentation for more information.

Discorder.list_guild_stickers — Function

list_guild_stickers(client, guild_id) -> Vector{Sticker}

Make a GET request to /guilds/<guild_id>/stickers. See the Discord API documentation for more information.

Discorder.list_joined_private_archived_threads — Function

list_joined_private_archived_threads(client, channel_id; kwargs...) -> ArchivedThread

Make a GET request to /channels/<channel_id>/users/@me/threads/archived/private. See the Discord API documentation for more information.

Discorder.list_nitro_sticker_packs — Function

list_nitro_sticker_packs(client) -> Vector{Sticker}

Make a GET request to /sticker-packs. See the Discord API documentation for more information.

Discorder.list_private_archived_threads — Function

list_private_archived_threads(client, channel_id; kwargs...) -> ArchivedThread

Make a GET request to /channels/<channel_id>/threads/archived/private. See the Discord API documentation for more information.

Discorder.list_public_archived_threads — Function

list_public_archived_threads(client, channel_id; kwargs...) -> ArchivedThread

Make a GET request to /channels/<channel_id>/threads/archived/public. See the Discord API documentation for more information.

Discorder.list_thread_members — Function

list_thread_members(client, channel_id) -> Vector{ThreadMember}

Make a GET request to /channels/<channel_id>/thread-members. See the Discord API documentation for more information.

Discorder.list_voice_regions — Function

list_voice_regions(client) -> Vector{VoiceRegion}

Make a GET request to /voice/regions. See the Discord API documentation for more information.

Discorder.modify_channel — Function

modify_channel(client, channel_id; kwargs...) -> DiscordChannel

Make a PATCH request to /channels/<channel_id>. See the Discord API documentation for more information.

Discorder.modify_current_member — Function

modify_current_member(client, guild_id; kwargs...) -> Nothing

Make a PATCH request to /guilds/<guild_id>/members/@me. See the Discord API documentation for more information.

Discorder.modify_current_user — Function

modify_current_user(client; kwargs...) -> User

Make a PATCH request to /users/@me. See the Discord API documentation for more information.

Discorder.modify_current_user_nick — Function

modify_current_user_nick(client, guild_id; kwargs...) -> UserNickChange

Make a PATCH request to /guilds/<guild_id>/members/@me/nick. See the Discord API documentation for more information.

Discorder.modify_current_user_voice_state — Function

modify_current_user_voice_state(client, guild_id; kwargs...) -> Nothing

Make a PATCH request to /guilds/<guild_id>/voice-states/@me. See the Discord API documentation for more information.

Discorder.modify_guild — Function

modify_guild(client, guild_id; kwargs...) -> Guild

Make a PATCH request to /guilds/<guild_id>. See the Discord API documentation for more information.

Discorder.modify_guild_channel_positions — Function

modify_guild_channel_positions(client, guild_id; positions) -> Nothing

Make a PATCH request to /guilds/<guild_id>/channels. See the Discord API documentation for more information.

Discorder.modify_guild_emoji — Function

modify_guild_emoji(client, guild_id, emoji; kwargs...) -> Emoji

Make a PATCH request to /guilds/<guild_id>/emojis/<emoji>. See the Discord API documentation for more information.

Discorder.modify_guild_member — Function

modify_guild_member(client, guild_id, user_id; kwargs...) -> Nothing

Make a PATCH request to /guilds/<guild_id>/members/<user_id>. See the Discord API documentation for more information.

Discorder.modify_guild_role — Function

modify_guild_role(client, guild_id, role_id; kwargs...) -> Role

Make a PATCH request to /guilds/<guild_id>/roles/<role_id>. See the Discord API documentation for more information.

Discorder.modify_guild_role_positions — Function

modify_guild_role_positions(client, guild_id; positions) -> Vector{Role}

Make a PATCH request to /guilds/<guild_id>/roles. See the Discord API documentation for more information.

Discorder.modify_guild_scheduled_event — Function

modify_guild_scheduled_event(client, guild_id, guild_scheduled_event_id; kwargs...) -> GuildScheduledEvent

Make a PATCH request to /guilds/<guild_id>/scheduled-events/<guild_scheduled_event_id>. See the Discord API documentation for more information.

Discorder.modify_guild_sticker — Function

modify_guild_sticker(client, guild_id, sticker_id; kwargs...) -> Sticker

Make a PATCH request to /guilds/<guild_id>/stickers/<sticker_id>. See the Discord API documentation for more information.

Discorder.modify_guild_template — Function

modify_guild_template(client, guild_id, template_code; kwargs...) -> GuildTemplate

Make a PATCH request to /guilds/<guild_id>/templates/<template_code>. See the Discord API documentation for more information.

Discorder.modify_guild_welcome_screen — Function

modify_guild_welcome_screen(client, guild_id; kwargs...) -> WelcomeScreen

Make a PATCH request to /guilds/<guild_id>/welcome-screen. See the Discord API documentation for more information.

Discorder.modify_guild_widget — Function

modify_guild_widget(client, guild_id; kwargs...) -> GuildWidget

Make a PATCH request to /guilds/<guild_id>/widget. See the Discord API documentation for more information.

Discorder.modify_stage_instance — Function

modify_stage_instance(client; kwargs...) -> StageInstance

Make a PATCH request to /stage-instances/{channel.id}. See the Discord API documentation for more information.

Discorder.modify_user_voice_state — Function

modify_user_voice_state(client, guild_id, user_id; kwargs...) -> Nothing

Make a PATCH request to /guilds/<guild_id>/voice-states/<user_id>. See the Discord API documentation for more information.

Discorder.modify_webhook — Function

modify_webhook(client, webhook_id; kwargs...) -> Webhook

Make a PATCH request to /webhooks/<webhook_id>. See the Discord API documentation for more information.

Discorder.modify_webhook_with_token — Function

modify_webhook_with_token(client, webhook_id, webhook_token; kwargs...) -> Webhook

Make a PATCH request to /webhooks/<webhook_id>/<webhook_token>. See the Discord API documentation for more information.

Discorder.pin_message — Function

pin_message(client, channel_id, message_id) -> Nothing

Make a PUT request to /channels/<channel_id>/pins/<message_id>. See the Discord API documentation for more information.

Discorder.publish — Method

publish(publisher::AbstractEventPublisher, event::Event)

Publish an event via the provided event publisher.

Discorder.remove_group_dm_recipient — Function

remove_group_dm_recipient(client, channel_id, user_id) -> Nothing

Make a DELETE request to /channels/<channel_id>/recipients/<user_id>. See the Discord API documentation for more information.

Discorder.remove_guild_ban — Function

remove_guild_ban(client, guild_id, user_id) -> Nothing

Make a DELETE request to /guilds/<guild_id>/bans/<user_id>. See the Discord API documentation for more information.

Discorder.remove_guild_member — Function

remove_guild_member(client, guild_id, user_id) -> Nothing

Make a DELETE request to /guilds/<guild_id>/members/<user_id>. See the Discord API documentation for more information.

Discorder.remove_guild_member_role — Function

remove_guild_member_role(client, guild_id, user_id, role_id) -> Nothing

Make a DELETE request to /guilds/<guild_id>/members/<user_id>/roles/<role_id>. See the Discord API documentation for more information.

Discorder.remove_thread_member — Function

remove_thread_member(client, channel_id, user_id) -> Nothing

Make a DELETE request to /channels/<channel_id>/thread-members/<user_id>. See the Discord API documentation for more information.

Discorder.safe_parse_json — Method

safe_parse_json(tracker::GatewayTracker, json::AbstractString, T::DataType)

Parse a JSON string into an expected type T, which is configured using StructType traits. Returns nothing if the string cannot be parsed for some reasons. Throw exception only during fail_on_error mode.

Discorder.safe_wait — Method

safe_wait(task::Optional{Task})

Wait for a task for finish synchronously. Unlike wait, it is meant to be safe and never throw, except during fail_on_error mode. Always return nothing.

Discorder.sanitize — Method

sanitize(s::AbstractString)

Remove sensitive information such as Bot token.

Discorder.search_guild_members — Function

search_guild_members(client, guild_id; kwargs...) -> Vector{GuildMember}

Make a GET request to /guilds/<guild_id>/members/search. See the Discord API documentation for more information.

Discorder.serve — Method

serve(;
    client::BotClient=BotClient(),
    tracker_ref=Ref{GatewayTracker}(),
    config_file_path::Optional{AbstractString}=nothing,
)

Run control plane in a loop so that we can actually auto-recover when bad things happen.

Discorder.should_trigger — Function

Return a vector of arguments that is passed to the user function when triggered, or nothing if the trigger should not be fired.

Discorder.show_error — Method

show_error(ex::Exception)

Display an error in the log with backtrace. This is usefulf for debugging purpose.

Discorder.start_control_plane — Method

start_control_plane(client::BotClient, config)

Start new control plane and return a GatewayTracker object.

Discorder.start_thread_from_message — Function

start_thread_from_message(client, channel_id, message_id; kwargs...) -> DiscordChannel

Make a POST request to /channels/<channel_id>/messages/<message_id>/threads. See the Discord API documentation for more information.

Discorder.start_thread_without_message — Function

start_thread_without_message(client, channel_id; kwargs...) -> DiscordChannel

Make a POST request to /channels/<channel_id>/threads. See the Discord API documentation for more information.

Discorder.sync_guild_template — Function

sync_guild_template(client, guild_id, template_code) -> GuildTemplate

Make a PUT request to /guilds/<guild_id>/templates/<template_code>. See the Discord API documentation for more information.

Discorder.trigger_typing_indicator — Function

trigger_typing_indicator(client, channel_id) -> Nothing

Make a POST request to /channels/<channel_id>/typing. See the Discord API documentation for more information.

Discorder.unpin_message — Function

unpin_message(client, channel_id, message_id) -> Nothing

Make a DELETE request to /channels/<channel_id>/pins/<message_id>. See the Discord API documentation for more information.

Discorder.waterfall — Method

waterfall(f, xs)

Apply function f to each element of collection xs iteratively and return the very first application where no exception was thrown. If every application throws exception then the last exception is re-thrown.

Discorder.@discord_object — Macro

Define a Discord resource.